Initial Commit
Simple script to capture the output of a command, and write it to a
logfile.  The existing file with the given name will be renamed to .old.
This will cause the AppFirst collector to think that the log file has
been "rotated", and therefore start reading from the beginning.  The
point of all this is to capture any command output to a log file, while
consuming neglible local disk space.

This script was designed to be run in a cron job.